# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /job-listings:
    get:
      operationId: listJobListings
      tags: [job-listings]
      summary: List active job listings
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/JobListing"
    post:
      operationId: createJobListing
      tags: [job-listings]
      summary: Create a job listing (admin)
      parameters:
        - in: header
          name: x-admin-password
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JobListingInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobListing"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /job-listings/{id}:
    put:
      operationId: updateJobListing
      tags: [job-listings]
      summary: Update a job listing (admin)
      parameters:
        - in: header
          name: x-admin-password
          required: true
          schema:
            type: string
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JobListingInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobListing"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      operationId: deleteJobListing
      tags: [job-listings]
      summary: Delete a job listing (admin)
      parameters:
        - in: header
          name: x-admin-password
          required: true
          schema:
            type: string
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"

  # ── Startup Applications ───────────────────────────────────
...
  /humanities/quiz/today:
    get:
      operationId: getTodayQuiz
      tags: [humanities-quizzes]
      summary: Get today's O/X quiz and attempt status
      parameters:
        - in: query
          name: sessionKey
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HumanitiesQuizTodayResponse"
  /humanities/quiz/attempt:
    post:
      operationId: submitQuizAttempt
      tags: [humanities-quizzes]
      summary: Submit quiz attempt
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HumanitiesQuizAttemptInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HumanitiesQuizAttempt"
        "409":
          description: Already attempted
  /humanities/quizzes:
    get:
      operationId: listHumanitiesQuizzes
      tags: [humanities-quizzes]
      summary: List all quizzes with stats (admin)
      parameters:
        - in: header
          name: x-admin-password
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/HumanitiesQuizWithStats"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      operationId: createHumanitiesQuiz
      tags: [humanities-quizzes]
      summary: Create quiz (admin)
      parameters:
        - in: header
          name: x-admin-password
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HumanitiesQuizInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HumanitiesQuiz"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /humanities/quizzes/{id}:
    put:
      operationId: updateHumanitiesQuiz
      tags: [humanities-quizzes]
      summary: Update quiz (admin)
      parameters:
        - in: header
          name: x-admin-password
          required: true
          schema:
            type: string
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HumanitiesQuizInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HumanitiesQuiz"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      operationId: deleteHumanitiesQuiz
      tags: [humanities-quizzes]
      summary: Delete quiz (admin)
      parameters:
        - in: header
          name: x-admin-password
          required: true
          schema:
            type: string
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"

  # ── Humanities Articles ───────────────────────────────────
  /humanities/articles:
    get:
      operationId: listHumanitiesArticles
      tags: [humanities-articles]
      summary: List active humanities articles
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/HumanitiesArticle"
    post:
      operationId: createHumanitiesArticle
      tags: [humanities-articles]
      summary: Create article (admin)
      parameters:
        - in: header
          name: x-admin-password
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HumanitiesArticleInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HumanitiesArticle"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /humanities/articles/{id}:
    get:
      operationId: getHumanitiesArticle
      tags: [humanities-articles]
      summary: Get article detail
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HumanitiesArticle"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      operationId: updateHumanitiesArticle
      tags: [humanities-articles]
      summary: Update article (admin)
      parameters:
        - in: header
          name: x-admin-password
          required: true
          schema:
            type: string
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HumanitiesArticleInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HumanitiesArticle"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      operationId: deleteHumanitiesArticle
      tags: [humanities-articles]
      summary: Delete article (admin)
      parameters:
        - in: header
          name: x-admin-password
          required: true
          schema:
            type: string
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
components:
  responses:
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
  schemas:
...
    ApiError:
      type: object
      properties:
        error:
          type: string
      required: [error]

    # ── Mentor Application ─────────────────────────────────
...
    LearningItem:
      type: object
      properties:
        title:
          type: string
        content:
          type: string
      required: [title, content]
    JobListing:
      type: object
      properties:
        id:
          type: integer
        category:
          type: string
        title:
          type: string
        shortDescription:
          type: string
        imageUrl:
          type: ["string", "null"]
        isActive:
          type: boolean
        displayOrder:
          type: integer
        learning:
          type: array
          items:
            $ref: "#/components/schemas/LearningItem"
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - category
        - title
        - shortDescription
        - imageUrl
        - isActive
        - displayOrder
        - learning
        - createdAt
    JobListingInput:
      type: object
      properties:
        category:
          type: string
          minLength: 1
        title:
          type: string
          minLength: 1
        shortDescription:
          type: string
        imageUrl:
          type: ["string", "null"]
        isActive:
          type: boolean
        displayOrder:
          type: integer
        learning:
          type: array
          items:
            $ref: "#/components/schemas/LearningItem"
      required: [category, title]

    # ── Startup Application ────────────────────────────────
...
    HumanitiesQuiz:
      type: object
      properties:
        id:
          type: integer
        question:
          type: string
        answer:
          type: boolean
        explanation:
          type: string
        scheduledDate:
          type: ["string", "null"]
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required: [id, question, answer, explanation, scheduledDate, isActive, createdAt]
    HumanitiesQuizWithStats:
      allOf:
        - $ref: "#/components/schemas/HumanitiesQuiz"
        - type: object
          properties:
            participantCount:
              type: integer
            correctCount:
              type: integer
            correctRate:
              type: integer
          required: [participantCount, correctCount, correctRate]
    HumanitiesQuizInput:
      type: object
      properties:
        question:
          type: string
          minLength: 1
        answer:
          type: boolean
        explanation:
          type: string
        scheduledDate:
          type: ["string", "null"]
        isActive:
          type: boolean
      required: [question, answer]
    HumanitiesQuizTodayResponse:
      type: object
      properties:
        quiz:
          oneOf:
            - $ref: "#/components/schemas/HumanitiesQuiz"
            - type: "null"
        attempt:
          type: object
          nullable: true
          properties:
            id:
              type: integer
            quizId:
              type: integer
            sessionKey:
              type: string
            isCorrect:
              type: boolean
            attemptedAt:
              type: string
              format: date-time
      required: [quiz, attempt]
    HumanitiesQuizAttemptInput:
      type: object
      properties:
        quizId:
          type: integer
        sessionKey:
          type: string
        isCorrect:
          type: boolean
      required: [quizId, sessionKey, isCorrect]
    HumanitiesQuizAttempt:
      type: object
      properties:
        id:
          type: integer
        quizId:
          type: integer
        sessionKey:
          type: string
        isCorrect:
          type: boolean
        attemptedAt:
          type: string
          format: date-time
      required: [id, quizId, sessionKey, isCorrect, attemptedAt]

    # ── Humanities Article ────────────────────────────────────
    HumanitiesArticle:
      type: object
      properties:
        id:
          type: integer
        category:
          type: string
        title:
          type: string
        content:
          type: string
        authorName:
          type: string
        imageUrl:
          type: string
        isActive:
          type: boolean
        displayOrder:
          type: integer
        createdAt:
          type: string
          format: date-time
      required: [id, category, title, content, authorName, imageUrl, isActive, displayOrder, createdAt]
    HumanitiesArticleInput:
      type: object
      properties:
        category:
          type: string
          minLength: 1
        title:
          type: string
          minLength: 1
        content:
          type: string
        authorName:
          type: string
        imageUrl:
          type: string
        isActive:
          type: boolean
        displayOrder:
          type: integer
      required: [category, title]