# lib/api-zod/src/generated/types/humanitiesQuiz.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface HumanitiesQuiz {
  id: number;
  question: string;
  answer: boolean;
  explanation: string;
  /** @nullable */
  scheduledDate: string | null;
  isActive: boolean;
  createdAt: Date;
}