# lib/api-zod/src/generated/types/humanitiesQuizAttemptInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface HumanitiesQuizAttemptInput {
  quizId: number;
  sessionKey: string;
  isCorrect: boolean;
}