# lib/api-zod/src/generated/types/humanitiesQuizWithStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { HumanitiesQuiz } from "./humanitiesQuiz";

export type HumanitiesQuizWithStats = HumanitiesQuiz & {
  participantCount: number;
  correctCount: number;
  correctRate: number;
};